models: add a few extra user tests

+++ models/user_test.go
@@ -54,10 +54,12 @@ func TestUserManager_GetPrivateIPFSNetworksForUSer(t *testing.T) {
 	)
 
 	tests := []struct {
-		name string
-		args args
+		name    string
+		args    args
+		wantErr bool
 	}{
-		{"Success", args{username, email, "password123"}},
+		{"Success", args{username, email, "password123"}, false},
+		{"Failure", args{"notarealuser", "notarealemail", "password123"}, true},
 	}
 
 	for _, tt := range tests {
@@ -67,8 +69,8 @@ func TestUserManager_GetPrivateIPFSNetworksForUSer(t *testing.T) {
 				t.Fatal(err)
 			}
 			defer um.DB.Delete(user)
-			if _, err := um.GetPrivateIPFSNetworksForUser(tt.args.userName); err != nil {
-				t.Fatal(err)
+			if _, err := um.GetPrivateIPFSNetworksForUser(tt.args.userName); (err != nil) != tt.wantErr {
+				t.Fatalf("GetPrivateIPFSNetworksForUser() wantErr = %v, error = %v", tt.wantErr, err.Error())
 			}
 		})
 	}
@@ -93,28 +95,25 @@ func TestUserManager_CheckIfUserHasAccessToNetwork(t *testing.T) {
 	)
 
 	tests := []struct {
-		name string
-		args args
+		name    string
+		args    args
+		wantErr bool
 	}{
-		{"Success", args{username, email, "password123"}},
+		{"Success", args{username, email, "password123"}, false},
+		{"Failure", args{"notarealuser", "notarealemail", "password123"}, true},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			user, err := um.NewUserAccount(tt.args.userName, tt.args.password, tt.args.email)
-			if err != nil {
-				t.Fatal(err)
+			if tt.name == "Success" {
+				user, err := um.NewUserAccount(tt.args.userName, tt.args.password, tt.args.email)
+				if err != nil {
+					t.Fatal(err)
+				}
+				defer um.DB.Delete(user)
 			}
-			defer um.DB.Delete(user)
-			access, err := um.CheckIfUserHasAccessToNetwork(
-				tt.args.userName,
-				testNetwork,
-			)
-			if err != nil {
-				t.Fatal("err")
-			}
-			if access {
-				t.Fatal("access to non existent network, this should not happen")
+			if _, err := um.CheckIfUserHasAccessToNetwork(tt.args.userName, testNetwork); (err != nil) != tt.wantErr {
+				t.Fatalf("CheckIfUserHasAccessToNetwork() wantErr = %v, error = %v", tt.wantErr, err.Error())
 			}
 		})
 	}
